Implmented Parcelable
I modified the java files to implement Parcelable so that they could be
passed between activities. The old way was to create a static variable
and reference it from another activity. The general concesus on the
internet forms seemed to indicate that this is the best way to pass
objects between activities.