Added functionality to post button
The post button now adds whatever it typed into the edit text next to it
to to the list of a picture's comments. These new comments are not sent
back to the list in the main menu, however, so the new comments will be
gone if the user leaves and comes back.